// data/text-data/recovery-password.enum.ts
export enum RecoveryPasswordTextData {
    STEP_TWO_MESSAGE = "We have sent an email with a 6-digit verification code",
    ERROR_ID_MESSAGE = "Invalid characters. ID no. must be at least 6 and not more than 20 characters",
    ERROR_NO_USER_FOUND = "No user found with this ID no. Please check your ID no. or contact the bank",
    TITLE = "Reset password",
    ENTER_ID = "Enter ID no.",
    STEP1_TEXT = "Enter ID no.",
    PLACEHOLDER = "Enter ID no.",
    STEP2_TEXT = "Enter code",
    STEP2_LABEL = "Enter code",
    STEP2_PLACEHOLDER = "Enter code",
    STEP2_RESENGING = "Resending the code will be available in",
    STEP3_TEXT = "Create new password",
    MESSAGE_ABOUT_CODE = "One-time verification code will be sent to your email address",
    STEP2_CODE_IS_VALID = "The code is valid for 10 minutes. Do not tell it to anyone. Make sure the email you received is from Polybank",
    INVALID_CODE = "Invalid code, you have 2 attempts left",
    PASSWORD_PLACEHOLDER = "Enter new password",
    CONFIRM_PASSWORD_PLACEHOLDER = "Confirm password",
    WRONG_PASSWORD_MESSAGE = "Password must be at least 8, no longer than 20 characters and contain: lowercase/capital Latin letters, numbers",
    STEP4_SUCCESS_TITLE = "Congrats!",
    STEP4_PASSWORD_CHANGED = "Password changed successfully.",
    STEP4_YOU_CAN_LOGIN = "Now you can log in to your account",
}